executer, trex-client Move trex modules to new package
